Refactor: Seperate Game.Player.cs
